rgb/hackathon/bcf2026prep/payment_microservice/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
pika>=1.3.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
prometheus-client>=0.19.0
